ISSUE #2944 AVATARS_COL_NAME moved to useres.constants
