+++ backend/src/v5/processors/teamspaces/teamspaces.js
@@ -19,7 +19,7 @@ const { addDefaultJobs, assignUserToJob, getJobsToUsers } = require('../../model
 const { createTeamspaceRole, grantTeamspaceRoleToUser, revokeTeamspaceRoleFromUser } = require('../../models/roles');
 const { createTeamspaceSettings, getMembersInfo, removeUserFromAdminPrivilege } = require('../../models/teamspaces');
 const { getAccessibleTeamspaces, grantAdminToUser } = require('../../models/users');
-const { AVATARS_COL_NAME } = require('../../models/fileRefs.constants');
+const { AVATARS_COL_NAME } = require('../../models/users.constants');
 const { DEFAULT_OWNER_JOB } = require('../../models/jobs.constants');
 const { getFileAsStream } = require('../../services/filesManager');
 const { isTeamspaceAdmin } = require('../../utils/permissions/permissions');
